# Optimize min/max queries with table statistics