--- a/datafusion/src/optimizer/aggregate_statistics.rs
+++ b/datafusion/src/optimizer/aggregate_statistics.rs
@@ -70,20 +70,22 @@ impl OptimizerRule for AggregateStatistics {
                         if let Some(column_statistics) =
                             source.statistics().column_statistics
                         {
-                            for (i, field) in fields.iter().enumerate() {
-                                if let Some(max_value) =
-                                    column_statistics[i].max_value.clone()
-                                {
-                                    let max_key =
-                                        format!("{}.{}", table_name, field.name());
-                                    max_values.insert(max_key, max_value);
-                                }
-                                if let Some(min_value) =
-                                    column_statistics[i].min_value.clone()
-                                {
-                                    let min_key =
-                                        format!("{}.{}", table_name, field.name());
-                                    min_values.insert(min_key, min_value);
+                            if fields.len() == column_statistics.len() {
+                                for (i, field) in fields.iter().enumerate() {
+                                    if let Some(max_value) =
+                                        column_statistics[i].max_value.clone()
+                                    {
+                                        let max_key =
+                                            format!("{}.{}", table_name, field.name());
+                                        max_values.insert(max_key, max_value);
+                                    }
+                                    if let Some(min_value) =
+                                        column_statistics[i].min_value.clone()
+                                    {
+                                        let min_key =
+                                            format!("{}.{}", table_name, field.name());
+                                        min_values.insert(min_key, min_value);
+                                    }
                                 }
                             }
                         }
@@ -240,13 +242,18 @@ mod tests {
     use crate::logical_plan::LogicalPlan;
     use crate::optimizer::aggregate_statistics::AggregateStatistics;
     use crate::optimizer::optimizer::OptimizerRule;
+    use crate::scalar::ScalarValue;
     use crate::{
-        datasource::{datasource::Statistics, TableProvider},
+        datasource::{
+            datasource::{ColumnStatistics, Statistics},
+            TableProvider,
+        },
         logical_plan::Expr,
     };
 
     struct TestTableProvider {
         num_rows: usize,
+        column_statistics: Vec<ColumnStatistics>,
         is_exact: bool,
     }
 
@@ -267,11 +274,11 @@ mod tests {
         ) -> Result<std::sync::Arc<dyn crate::physical_plan::ExecutionPlan>> {
             unimplemented!()
         }
-        fn statistics(&self) -> crate::datasource::datasource::Statistics {
+        fn statistics(&self) -> Statistics {
             Statistics {
                 num_rows: Some(self.num_rows),
                 total_byte_size: None,
-                column_statistics: None,
+                column_statistics: Some(self.column_statistics.clone()),
             }
         }
         fn has_exact_statistics(&self) -> bool {
@@ -287,6 +294,7 @@ mod tests {
             "test",
             Arc::new(TestTableProvider {
                 num_rows: 100,
+                column_statistics: Vec::new(),
                 is_exact: true,
             }),
         )
@@ -312,6 +320,7 @@ mod tests {
             "test",
             Arc::new(TestTableProvider {
                 num_rows: 100,
+                column_statistics: Vec::new(),
                 is_exact: false,
             }),
         )
@@ -337,6 +346,7 @@ mod tests {
             "test",
             Arc::new(TestTableProvider {
                 num_rows: 100,
+                column_statistics: Vec::new(),
                 is_exact: true,
             }),
         )
@@ -363,6 +373,7 @@ mod tests {
             "test",
             Arc::new(TestTableProvider {
                 num_rows: 100,
+                column_statistics: Vec::new(),
                 is_exact: true,
             }),
         )
@@ -388,6 +399,7 @@ mod tests {
             "test",
             Arc::new(TestTableProvider {
                 num_rows: 100,
+                column_statistics: Vec::new(),
                 is_exact: true,
             }),
         )
@@ -401,6 +413,68 @@ mod tests {
             \n  Aggregate: groupBy=[[]], aggr=[[COUNT(UInt8(1))]]\
             \n    Filter: #test.a Lt Int64(5)\
             \n      TableScan: test projection=None";
+
+        assert_optimized_plan_eq(&plan, expected);
+        Ok(())
+    }
+
+    #[test]
+    fn optimize_max_min_using_statistics() -> Result<()> {
+        use crate::execution::context::ExecutionContext;
+        let mut ctx = ExecutionContext::new();
+
+        let column_statistic = ColumnStatistics {
+            null_count: None,
+            max_value: Some(ScalarValue::from(100_i64)),
+            min_value: Some(ScalarValue::from(1_i64)),
+            distinct_count: None,
+        };
+        let mut column_statistics = Vec::new();
+        column_statistics.push(column_statistic);
+
+        ctx.register_table(
+            "test",
+            Arc::new(TestTableProvider {
+                num_rows: 100,
+                column_statistics: column_statistics,
+                is_exact: true,
+            }),
+        )
+        .unwrap();
+
+        let plan = ctx
+            .create_logical_plan("select max(a), min(a) from test")
+            .unwrap();
+        let expected = "\
+            Projection: #MAX(test.a), #MIN(test.a)\
+            \n  Projection: Int64(100) AS MAX(a), Int64(1) AS MIN(a)\
+            \n    EmptyRelation";
+
+        assert_optimized_plan_eq(&plan, expected);
+        Ok(())
+    }
+
+    #[test]
+    fn optimize_max_min_not_using_statistics() -> Result<()> {
+        use crate::execution::context::ExecutionContext;
+        let mut ctx = ExecutionContext::new();
+        ctx.register_table(
+            "test",
+            Arc::new(TestTableProvider {
+                num_rows: 100,
+                column_statistics: Vec::new(),
+                is_exact: true,
+            }),
+        )
+        .unwrap();
+
+        let plan = ctx
+            .create_logical_plan("select max(a), min(a) from test")
+            .unwrap();
+        let expected = "\
+            Projection: #MAX(test.a), #MIN(test.a)\
+            \n  Aggregate: groupBy=[[]], aggr=[[MAX(#test.a), MIN(#test.a)]]\
+            \n    TableScan: test projection=None";
 
         assert_optimized_plan_eq(&plan, expected);
         Ok(())