# Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Closes #640.

 # Rationale for this change
<!--
 Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
 Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.  
-->
Speeding up max and min queries that can utilize the availability of the table column statistics.

On the TPC-H `lineitem` table we can see a 600x speedup:
```
> CREATE EXTERNAL TABLE X STORED AS PARQUET LOCATION './benchmarks/parquet/lineitem/';
> select max(l_orderkey) from X;
+-----------------+
| MAX(l_orderkey) |
+-----------------+
| 6000000         |
+-----------------+
1 row in set. Query took 0.004 seconds.
> select min(l_quantity) from X;
+-----------------+
| MIN(l_quantity) |
+-----------------+
| 1               |
+-----------------+
1 row in set. Query took 0.003 seconds.
```
Master:
```
> CREATE EXTERNAL TABLE X STORED AS PARQUET LOCATION './benchmarks/parquet/lineitem/';
> select max(l_orderkey) from X;
+-----------------+
| MAX(l_orderkey) |
+-----------------+
| 6000000         |
+-----------------+
1 row in set. Query took 2.589 seconds.
> select min(l_quantity) from X;
+-----------------+
| MIN(l_quantity) |
+-----------------+
| 1               |
+-----------------+
1 row in set. Query took 2.146 seconds.
```

# What changes are included in this PR?
<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

# Are there any user-facing changes?
<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!--
If there are any breaking changes to public APIs, please add the `api change` label.
-->