@b41sh 2021-07-27
> I think this is looking great -- thank you @b41sh -- do you think this PR is ready for more review (it is still marked as a draft which is why I am asking)
> 
> cc @Dandandan this is a pretty cool one

It's basically ok, please review the code and give me some advice, thanks. @alamb