@b41sh 2021-07-28
> Looks like there are a few compile errors to sort out, but after that is fixed I think this should be ready to merge

This is because the `has_min_max_set` of `arrow-rs 5.0.0` is a private function, we may have to wait until `arrow-rs 5.1.0` is released