@alamb 2021-07-28
> we may have to wait until arrow-rs 5.1.0 is released

FYI the planned to release date of arrow-rs 5.1.0 is this weekend / early next week