@b41sh 2021-08-05
arrow-rs 5.1 has been released, can we test it? @alamb